Fixed Fast API server code errors
