+++ fast_api_server/constants/constants.py
